refactor(converter): split converter.go (1713 → 234 lines) by concern
converter.go mixed five separable concerns. Each is now its own file:

  types.go            errors, parser constants, Config/Device/etc. types, Parser
  parse_device.go     parseDevice + parseCapturePlayback + per-field helpers
  parse_protocols.go  SNMP/NetBIOS/ICMP/ICMPv6/DHCP/DNS/TTL parsers + merge helpers
  helpers.go          line/string utilities (extractValue, formatMAC, parseAddMib, ...)
  converter.go (slim) ConvertFile, Parse, LoadYAMLConfig*, ValidateConfig, PrintSummary

go test ./internal/converter/... passes; golangci-lint clean.
